example_04/requirements.txt
fastapi
uvicorn
pillow
python-multipart
pydantic[email]
